Disassociate actor executions from dataflow object preloads
Previously, actor executions would edit the 'preload' fields of the
dataflow graph itself. This meant that, at the end of the 'SymbExec'
algorithm, the dataflow graph would have a different number of tokens
from when it started. Use the 'currentTokens' field in the State
object instead to make token tracking separate unreliant on the
dataflow graph.