Fix bug in storage dependency detection for graphs with multiple SCC
Previously, storage dependency detection function would operate on SCC
of the main graph. This led to a bug where the maximum throughput of
the graph was always returned and not all storage dependencies were
correctly identified. This commit makes it such that the storage
dependency detection function considers the whole graph in its
execution, thereby fixing the bug.