Update cumulative time logs
Update cumulative time to be the time difference between the start of
the DSE algorithm and after the checklist has been populated.

Previously, the cumulative time stored was the sum of execution
times (time taken for throughput and critical cycle computation). This
did not account for the time taken to populate the checklist of
storage distributions as well as the time taken to minimize the
list of results.