Handle arithmetic operators with multiple input types
Remove check on the number of input types provided to an arithmetic
operator and print a warning message instead. Previously, if an
arithmetic operator had more than one input type (as specified by its
input port names), it would result in an "undefined" component. This
was an issue with operators with a "real" and "vect" inputs which, for
our purposes, we assume should result in a FP operator. This change
fixes that issue.