Update actor names during iterative evaluation
When bypassing delay actors or evaluating the result of various
operators, we now update the names of the resulting actors
accordingly. Before doing this, we would get errors when running VHDL
generation using the resulting SDFs as the names either wouldn't
reflect the correct number of input arguments or would contain the
name of bypassed delay actors.