Handle cases with multiple inputs from a single actor
Previously, it was assumed that there would only be a single edge
between each actor. This had an impact on the way the
'getConnectionNameFromComponents' function was
implemented. Nonetheless, it has been observed that this is not always
the case. The implementation now returns a vector of signal names
rather than a single string. The order of input signals is also
preserved via the use of an intermediate vector (originalOrder).