Update actor names when merging outputs
Previously, the SingleOutput algorithm did not account for the fact
that the VHDLGeneration algorithm utilised actor names to identify
connections (via getConnectionNameFromComponents). This led to a bug
where getConnectionNameFromComponents would return the wrong number of
connection names as it would not correctly identify when an operator
now had a newly generated Proj operator as an input actor. We now
update this accordingly to avoid this bug.

Another bug addressed is when the names of actors updated in previous
iterations of SingleOutput are part of the list of actors that are
supposed to be merged --- this led to instances where SingleOutput
would not be able to find the actors in subsequent iterations. The
list of actors are now provided as a single argument, which is then
stored as an array (verticesToMerge). If any actor in this array has
their name updated, it is similarly updated in the list.