Determine delay argument and delayed signal based on channel number
Use the numbering included in channel names to determine whether an
input edge belongs to an actor supplying a delay argument or the
signal to be delayed. Previously, we looked for an input actor with a
type that was an integer as the delay argument and simply assigned the
other input actor as the delayed signal --- this led to bugs when both
input actors were integers. It has been observed that the first
channel (from the SDF .xml file) is the delayed signal while the
second is the delay argument. Nonetheless, the ForInputEdges loop
iterates through the channels according to their IDs, which rarely
correspond to their listed order in the .xml file. Extracting the
numbers from their channel names is used as a workaround for this.