Cast constant value data types to FP when necessary
Check the destination components of each constant value component ---
set the constant value data type and binary representation to floating
point if the data type of the destination component is mixed (and thus
expecting floating point input types). This fixes a bug when
generating VHDL for the power operator as Faust tends to take in an
integer input as an exponent argument while the VHDL operator expects
both the base and the exponent inputs to be the same type.