Fix bug in top level circuit generation
Previously, we used an input/output count to track and name the
signals connected to the top-level input/output components. This
resulted in a bug where sometimes a component would be connected to
the wrong output actor (e.g. OUTPUT_1 instead of OUTPUT_0). We now
track input and output actor IDs in the VHDLComponent class to ensure
that we can appropriately match the connection to the input/output
actors.