Fix vertex and edge naming conflicts
There are two bug fixes in this commit. The first is the vertex naming
conflict bug. Vertex names were being updated to account for when
their source vertices were changed to an output
selector (e.g. v_argX_argY -> v_outputselector_argY) so that we could
continue to correctly identify the order of arguments for the vertex
when generating the VHDL implementation. Nonetheless, we didn't
account for instances where the vertex didn't include the order of
arguments, leading to instances where vertex names were being entirely
replaced with the output selector name. To fix this bug, we now look for an underscore
character before renaming. The second bug is in channel naming
conflicts. We previously used channel IDs and total edge counts to try
to uniquely name new channels added while retaining the
"channel_N_type" convention. This still led to conflicts due to the
current channel names being static and the chance for channels being
deleted (thereby vacating an ID/edge count that's already being
used). We now add the ID and the edge count to try to keep the name
unique.