Fix bug with wrongly renamed actor names
Previously, we would replace the occurances of merged actors with the
name of the output selectors. This is to preserve the naming
convention of "actor_arg1_arg2" and to update "arg1" and "arg2" if
"actor" is now connected to an output selector. However, we were
mistakenly replacing instances of "arg1DupN" with
"outputselectorDupN". This led to errors where we'd be looking for an
actor named "outputselectorDupN" that didn't exist. We've fixed this
by checking for the substring "Dup" after the actor we're supposed to
be replacing.