Remove actor output merging as a default
This will now be require the parameter "NORMALISE_OP_OUTPUTS" to be
triggered when running the MergeOperators algorithm. This is because
the Proj operator (which was used to collect and send outputs) will no
longer be required in a non-data-driven implementation. It is
therefore removed from the SDF model of the implementation.