Add workarounds for argument order renaming
Ran into issues with identifying and updating the right order of
arguments for actors who have arguments from the same source
actor. I've added workarounds for now, but this will need to be
explicitly addressed in the future. Also, we no longer used shared
outputs for actors with multiple outputs --- this was a workaround
that is no longer necessary since we've implemented output merging in
time-triggered implementations.