Isolate audio interface wrapper generation from VHDLGenerator
Remove most of the hardcoded VHDL code generation from the
VHDLGenerator class. The idea is to have derived classes from a base
VHDLComponent class that will have functions to instantiate the
appropriate parameters and generate its necessary VHDL code. This
will resemble the hierarchy of VHDL components in the generated VHDL
implementation.